SteamIntegration (50%)
Functions made for getting gameData based on steamID, and for converting appID to game name for storage in database.

Still to do is a way to iterate through all games and put the game name, and then store in the database per user.

Possible issues: There is still no logic for if a user is using a private profile.